pyproject.toml
[project]
name = "werewolf"
version = "0.1.0"
description = "A CUI-based simple werewolf game with `autogen`"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "langchain==0.1.10",
    "langchain-openai==0.0.8",
    "pyautogen==0.2.16",
    "python-dotenv>=1.0.1",
]
authors = [{ name = "hmasdev" }]
maintainers = [{ name = "hmasdev" }]
readme = "README.md"
license = { file = "LICENSE" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "autopep8>=2.3.1",
    "flake8>=7.1.1",
    "mypy>=1.11.2",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[tool.uv]
dev-dependencies = [
    "autopep8>=2.3.1",
    "flake8>=7.1.1",
    "mypy>=1.11.2",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
addopts = "--cov=werewolf --cov-report=term-missing -m 'not long' -m 'not integration'"
markers = [
    "integration: mark a test as an integration test",
    "long: mark a test as a long-running test",
]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ['autogen']
ignore_missing_imports = true
